extract stuff from main into app
